+++ src/server.js
@@ -182,7 +182,7 @@ var Router = require("./router");
           .then(function (content) {
             if (!router.isDynamic(src)) {
               var file = src
-                .replace(router.src_dir, '')
+                .replace(router.srcDir, '')
                 .replace(/\.html$/, '');
               var customDst = router.globalFilename(file, args);
               return saveTemplate(name, src, content, dst, ctx).then(function () {
@@ -218,16 +218,16 @@ var Router = require("./router");
     return sequence(renders, function (f) { return f(); }, ctx);
   }
 
-  function renderDir(src_dir, dst_dir, router, ctx) {
-    return logAndTime("render dir '" + src_dir + "'", function () {
-      return createDirs([dst_dir], ctx)
+  function renderDir(srcDir, dstDir, router, ctx) {
+    return logAndTime("render dir '" + srcDir + "'", function () {
+      return createDirs([dstDir], ctx)
         .then(function () {
-          return Q.ninvoke(fs, 'readdir', src_dir);
+          return Q.ninvoke(fs, 'readdir', srcDir);
         })
         .then(function (names) {
           return sequence(names, function (name) {
-            var src = src_dir + "/" + name;
-            var dst = dst_dir + "/" + name;
+            var src = srcDir + "/" + name;
+            var dst = dstDir + "/" + name;
             return Q
               .ninvoke(fs, 'lstat', src)
               .then(function (stats) {
@@ -251,7 +251,7 @@ var Router = require("./router");
     }, ctx);
   }
 
-  function renderDynamicCall(call, dst_dir, router, ctx) {
+  function renderDynamicCall(call, dstDir, router, ctx) {
     var src = router.srcForCall(call);
     var dst = router.globalFilenameForCall(call);
     return createPath(dst, ctx).then(function () {
@@ -259,16 +259,16 @@ var Router = require("./router");
     });
   }
 
-  function renderStackedCalls(router, dst_dir, ctx) {
+  function renderStackedCalls(router, dstDir, ctx) {
     var lastCalls = router.lastCalls();
     if (_.isEmpty(lastCalls)) {
       return Q();
     } else {
       return sequence(lastCalls, function (call) {
         router.generated(call);
-        return renderDynamicCall(call, dst_dir, router, ctx);
+        return renderDynamicCall(call, dstDir, router, ctx);
       }, ctx).then(function () {
-        return renderStackedCalls(router, dst_dir, ctx);
+        return renderStackedCalls(router, dstDir, ctx);
       } );
     }
   }
@@ -309,13 +309,13 @@ var Router = require("./router");
     }, ctx);
   }
 
-  function buildRouterForDir(src_dir, ctx) {
-    return logAndTime("Build router for dir '" + src_dir + "'", function () {
+  function buildRouterForDir(srcDir, ctx) {
+    return logAndTime("Build router for dir '" + srcDir + "'", function () {
       return Q
-        .ninvoke(fs, 'readdir', src_dir)
+        .ninvoke(fs, 'readdir', srcDir)
         .then(function (names) {
           return sequence(names, function (name) {
-            var src = src_dir + "/" + name;
+            var src = srcDir + "/" + name;
             return Q
               .ninvoke(fs, 'lstat', src)
               .then(function (stats) {
@@ -336,11 +336,11 @@ var Router = require("./router");
     }, ctx);
   }
 
-  function buildRouter(src_dir, dst_dir, ctx) {
-    return buildRouterForDir(src_dir, ctx).then(function (params) {
+  function buildRouter(srcDir, dstDir, ctx) {
+    return buildRouterForDir(srcDir, ctx).then(function (params) {
       return Router.create(params, {
-        src_dir: src_dir,
-        dst_dir: dst_dir,
+        srcDir: srcDir,
+        dstDir: dstDir,
         logger: ctx.logger
       });
     });
@@ -369,26 +369,26 @@ var Router = require("./router");
   function run(opts) {
     return Q.fcall(function () {
       var ctx = _.assign({logger: buildLogger(opts)}, opts, {
-        src_dir: opts.src_dir.replace(/\/$/, ''),
-        dst_dir: opts.dst_dir.replace(/\/$/, '')
+        srcDir: opts.srcDir.replace(/\/$/, ''),
+        dstDir: opts.dstDir.replace(/\/$/, '')
       });
       ctx.logger.info("ctx:", _.assign({}, ctx, {logger: '<LOGGER>'}));
       if (ctx.debug) Q.longStackSupport = true;
       return logAndTime("Generation", function () {
-        return createPaths([ctx.dst_dir], ctx)
+        return createPaths([ctx.dstDir], ctx)
           .then(function () {
-            return buildRouter(ctx.src_dir, ctx.dst_dir, ctx);
+            return buildRouter(ctx.srcDir, ctx.dstDir, ctx);
           })
           .then(function (router) {
-            return renderDir(ctx.src_dir, ctx.dst_dir, router, ctx)
+            return renderDir(ctx.srcDir, ctx.dstDir, router, ctx)
               .then(function (fromFiles) {
-                return renderStackedCalls(router, ctx.dst_dir, ctx)
+                return renderStackedCalls(router, ctx.dstDir, ctx)
                   .then(function (fromCalls) {
                     return _.compact(fromFiles.concat(fromCalls));
                   });
               })
               .then(function (res) {
-                return saveRouter(router, ctx.dst_dir, ctx)
+                return saveRouter(router, ctx.dstDir, ctx)
                   .thenResolve(res);
               });
           });